refactor: create AbstractSocket and ConfigMixin classes
allocate common logic to this classes
