+++ usefull_stuff/messanger/common/abstract_socket.py
@@ -1,0 +1,60 @@
+"""
+Module with AbstractSocket class
+"""
+import json
+
+from usefull_stuff.messanger.common.config_mixin import ConfigMixin
+
+
+class AbstractSocket(ConfigMixin):
+    """
+    Abstract class containing common logic for SocketType objects
+    """
+    def __init__(self):
+        self._config = self.get_config()
+
+    @staticmethod
+    def get_message(client, config: dict) -> dict:
+        """
+        Receives and decodes messages
+
+        Args:
+            client:
+            config: socket config
+
+        Raises:
+            ValueError: if was mistake during conversion
+        """
+        max_msg_length = config.get('max_bytes_for_msg')
+        encoding = config.get('encoding')
+
+        encoded_response = client.recv(max_msg_length)
+        if isinstance(encoded_response, bytes):
+            json_response = encoded_response.decode(encoding)
+            if isinstance(json_response, str):
+                response = json.loads(json_response)
+                if isinstance(response, dict):
+                    return response
+                raise ValueError
+            raise ValueError
+        raise ValueError
+
+    @staticmethod
+    def send_message(sock, message: dict, config: dict) -> None:
+        """
+        Encodes and sends a message
+
+        Args:
+            sock:
+            message:
+            config: socket config
+        Raises:
+            TypeError: if got wrong message
+        """
+        encoding = config.get('encoding')
+
+        if not isinstance(message, dict):
+            raise TypeError
+        js_message = json.dumps(message)
+        encoded_message = js_message.encode(encoding)
+        sock.send(encoded_message)
